Add more output to update command
