+++ Tests/System/UpdateCommand/UpdateToSpecificVersionTest.php
@@ -1,0 +1,60 @@
+<?php
+
+namespace Tests\System\UpdateCommand\UpdateToSpecificVersionTest;
+
+use Saeghe\FileManager\FileType\Json;
+use function Saeghe\FileManager\Directory\clean;
+use function Saeghe\FileManager\Resolver\root;
+use function Saeghe\FileManager\Resolver\realpath;
+use function Saeghe\TestRunner\Assertions\Boolean\assert_true;
+
+test(
+    title: 'it should update to the given version',
+    case: function () {
+        $output = shell_exec('php ' . root() . 'saeghe update git@github.com:saeghe/released-package.git --version=v1.0.5 --project=TestRequirements/Fixtures/EmptyProject');
+
+        $expected = <<<EOD
+\e[39mUpdating package git@github.com:saeghe/released-package.git to version v1.0.5...
+\e[39mSetting env credential...
+\e[39mLoading configs...
+\e[39mFinding package in configs...
+\e[39mSetting package version...
+\e[39mLoading package's meta...
+\e[39mDeleting package's files...
+\e[39mDetecting version hash...
+\e[39mDownloading the package with new version...
+\e[39mUpdating configs...
+\e[39mCommitting new configs...
+\e[92mPackage git@github.com:saeghe/released-package.git has been updated.\e[39m
+
+EOD;
+
+        assert_true($expected === $output, 'Output is not correct:' . PHP_EOL . $expected . PHP_EOL . $output);
+        assert_given_version_added('Package did not updated to given package version. ' . $output);
+    },
+    before: function () {
+        shell_exec('php ' . root() . 'saeghe init --project=TestRequirements/Fixtures/EmptyProject');
+        shell_exec('php ' . root() . 'saeghe add git@github.com:saeghe/released-package.git --version=v1.0.3 --project=TestRequirements/Fixtures/EmptyProject');
+    },
+    after: function () {
+        clean(realpath(root() . 'TestRequirements/Fixtures/EmptyProject'));
+    }
+);
+
+function assert_given_version_added($message)
+{
+    $config = Json\to_array(realpath(root() . 'TestRequirements/Fixtures/EmptyProject/saeghe.config.json'));
+    $meta = Json\to_array(realpath(root() . 'TestRequirements/Fixtures/EmptyProject/saeghe.config-lock.json'));
+
+    assert_true((
+        isset($config['packages']['git@github.com:saeghe/released-package.git'])
+        && 'v1.0.5' === $config['packages']['git@github.com:saeghe/released-package.git']
+        && isset($meta['packages']['git@github.com:saeghe/released-package.git'])
+        && 'v1.0.5' === $meta['packages']['git@github.com:saeghe/released-package.git']['version']
+        && 'saeghe' === $meta['packages']['git@github.com:saeghe/released-package.git']['owner']
+        && 'released-package' === $meta['packages']['git@github.com:saeghe/released-package.git']['repo']
+        && '5885e5f3ed26c2289ceb2eeea1f108f7fbc10c01' === $meta['packages']['git@github.com:saeghe/released-package.git']['hash']
+    ),
+        $message
+    );
+}
