+++ Tests/System/UpdateCommand/UpdateWhenGithubTokenIsSetTest.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace Tests\System\UpdateCommand\UpdateWhenGithubTokenIsSetTest;
+
+use Saeghe\FileManager\FileType\Json;
+use function Saeghe\FileManager\Directory\clean;
+use function Saeghe\FileManager\Resolver\root;
+use function Saeghe\FileManager\Resolver\realpath;
+use function Saeghe\Saeghe\Providers\GitHub\github_token;
+use function Saeghe\TestRunner\Assertions\Boolean\assert_true;
+use const Saeghe\Saeghe\Providers\GitHub\GITHUB_DOMAIN;
+
+
+test(
+    title: 'it should not show error message if the credential file is not exist and GITHUB_TOKEN is set',
+    case: function () {
+        $output = shell_exec('php ' . root() . 'saeghe update git@github.com:saeghe/released-package.git --project=TestRequirements/Fixtures/EmptyProject');
+
+        $expected = <<<EOD
+\e[39mUpdating package git@github.com:saeghe/released-package.git to latest version...
+\e[39mSetting env credential...
+\e[39mLoading configs...
+\e[39mFinding package in configs...
+\e[39mSetting package version...
+\e[39mLoading package's meta...
+\e[39mDeleting package's files...
+\e[39mDetecting version hash...
+\e[39mDownloading the package with new version...
+\e[39mUpdating configs...
+\e[39mCommitting new configs...
+\e[92mPackage git@github.com:saeghe/released-package.git has been updated.\e[39m
+
+EOD;
+
+        assert_true($expected === $output, 'Output is not correct:' . PHP_EOL . $expected . PHP_EOL . $output);
+    },
+    before: function () {
+        shell_exec('php ' . root() . 'saeghe init --project=TestRequirements/Fixtures/EmptyProject');
+        shell_exec('php ' . root() . 'saeghe add git@github.com:saeghe/released-package.git --version=v1.0.3 --project=TestRequirements/Fixtures/EmptyProject');
+        $credential = Json\to_array(root() . 'credentials.json');
+        github_token($credential[GITHUB_DOMAIN]['token']);
+        rename(root() . 'credentials.json', root() . 'credentials.json.back');
+    },
+    after: function () {
+        clean(realpath(root() . 'TestRequirements/Fixtures/EmptyProject'));
+        rename(root() . 'credentials.json.back', root() . 'credentials.json');
+    }
+);
