--- Tests/System/UpdateCommandTest.php
@@ -1,170 +1,0 @@
-<?php
-
-namespace Tests\System\UpdateCommandTest;
-
-use Saeghe\FileManager\FileType\Json;
-use function Saeghe\Cli\IO\Write\assert_error;
-use function Saeghe\Cli\IO\Write\assert_success;
-use function Saeghe\FileManager\Directory\clean;
-use function Saeghe\FileManager\Resolver\root;
-use function Saeghe\FileManager\Resolver\realpath;
-use function Saeghe\Saeghe\Providers\GitHub\github_token;
-use function Saeghe\TestRunner\Assertions\Boolean\assert_true;
-use const Saeghe\Saeghe\Providers\GitHub\GITHUB_DOMAIN;
-
-test(
-    title: 'it should show error message if the credential file is not exist and there is no GITHUB_TOKEN',
-    case: function () {
-        $output = shell_exec('php ' . root() . 'saeghe update git@github.com:saeghe/released-package.git --project=TestRequirements/Fixtures/EmptyProject');
-
-        assert_error('There is no credential file. Please use the `credential` command to add your token.', $output);
-    },
-    before: function () {
-        shell_exec('php ' . root() . 'saeghe init --project=TestRequirements/Fixtures/EmptyProject');
-        shell_exec('php ' . root() . 'saeghe add git@github.com:saeghe/released-package.git --version=v1.0.3 --project=TestRequirements/Fixtures/EmptyProject');
-        rename(root() . 'credentials.json', root() . 'credentials.json.back');
-        github_token('');
-    },
-    after: function () {
-        clean(realpath(root() . 'TestRequirements/Fixtures/EmptyProject'));
-        rename(root() . 'credentials.json.back', root() . 'credentials.json');
-    }
-);
-
-test(
-    title: 'it should not show error message if the credential file is not exist and GITHUB_TOKEN is set',
-    case: function () {
-        $output = shell_exec('php ' . root() . 'saeghe update git@github.com:saeghe/released-package.git --project=TestRequirements/Fixtures/EmptyProject');
-
-        assert_success('Package git@github.com:saeghe/released-package.git has been updated.', $output);
-    },
-    before: function () {
-        shell_exec('php ' . root() . 'saeghe init --project=TestRequirements/Fixtures/EmptyProject');
-        shell_exec('php ' . root() . 'saeghe add git@github.com:saeghe/released-package.git --version=v1.0.3 --project=TestRequirements/Fixtures/EmptyProject');
-        $credential = Json\to_array(root() . 'credentials.json');
-        github_token($credential[GITHUB_DOMAIN]['token']);
-        rename(root() . 'credentials.json', root() . 'credentials.json.back');
-    },
-    after: function () {
-        clean(realpath(root() . 'TestRequirements/Fixtures/EmptyProject'));
-        rename(root() . 'credentials.json.back', root() . 'credentials.json');
-    }
-);
-
-test(
-    title: 'it should show error message when token is invalid',
-    case: function () {
-        $output = shell_exec('php ' . root() . 'saeghe update git@github.com:saeghe/released-package.git --project=TestRequirements/Fixtures/EmptyProject');
-
-        assert_error(
-            'The GitHub token is not valid. Either, you didn\'t set one yet, or it is not valid. Please use the `credential` command to set a valid token.',
-            $output
-        );
-    },
-    before: function () {
-        shell_exec('php ' . root() . 'saeghe init --project=TestRequirements/Fixtures/EmptyProject');
-        shell_exec('php ' . root() . 'saeghe add git@github.com:saeghe/released-package.git --version=v1.0.3 --project=TestRequirements/Fixtures/EmptyProject');
-        rename(root() . 'credentials.json', root() . 'credentials.json.back');
-        shell_exec('php ' . root() . 'saeghe credential github.com not-valid');
-        github_token('');
-    },
-    after: function () {
-        clean(realpath(root() . 'TestRequirements/Fixtures/EmptyProject'));
-        rename(root() . 'credentials.json.back', root() . 'credentials.json');
-    }
-);
-
-test(
-    title: 'it should update to the latest version',
-    case: function () {
-        $output = shell_exec('php ' . root() . 'saeghe update git@github.com:saeghe/released-package.git --project=TestRequirements/Fixtures/EmptyProject');
-
-        assert_success('Package git@github.com:saeghe/released-package.git has been updated.', $output);
-        assert_version_upgraded_in_config_file($output);
-        assert_meta_updated($output);
-    },
-    before: function () {
-        shell_exec('php ' . root() . 'saeghe init --project=TestRequirements/Fixtures/EmptyProject');
-        shell_exec('php ' . root() . 'saeghe add git@github.com:saeghe/released-package.git --version=v1.0.3 --project=TestRequirements/Fixtures/EmptyProject');
-    },
-    after: function () {
-        clean(realpath(root() . 'TestRequirements/Fixtures/EmptyProject'));
-    }
-);
-
-test(
-    title: 'it should show error message when package does not exists',
-    case: function () {
-        $output = shell_exec('php ' . root() . 'saeghe update git@github.com:saeghe/released-package.git --project=TestRequirements/Fixtures/EmptyProject');
-
-        assert_error('Package git@github.com:saeghe/released-package.git does not found in your project!', $output);
-    },
-    before: function () {
-        shell_exec('php ' . root() . 'saeghe init --project=TestRequirements/Fixtures/EmptyProject');
-        shell_exec('php ' . root() . 'saeghe add git@github.com:saeghe/simple-package.git --project=TestRequirements/Fixtures/EmptyProject');
-    },
-    after: function () {
-        clean(realpath(root() . 'TestRequirements/Fixtures/EmptyProject'));
-    }
-);
-
-test(
-    title: 'it should update to the given version',
-    case: function () {
-        $output = shell_exec('php ' . root() . 'saeghe update git@github.com:saeghe/released-package.git --version=v1.0.5 --project=TestRequirements/Fixtures/EmptyProject');
-
-        assert_given_version_added('Package did not updated to given package version. ' . $output);
-    },
-    before: function () {
-        shell_exec('php ' . root() . 'saeghe init --project=TestRequirements/Fixtures/EmptyProject');
-        shell_exec('php ' . root() . 'saeghe add git@github.com:saeghe/released-package.git --version=v1.0.3 --project=TestRequirements/Fixtures/EmptyProject');
-    },
-    after: function () {
-        clean(realpath(root() . 'TestRequirements/Fixtures/EmptyProject'));
-    }
-);
-
-function assert_version_upgraded_in_config_file($message)
-{
-    $config = Json\to_array(realpath(root() . 'TestRequirements/Fixtures/EmptyProject/saeghe.config.json'));
-
-    assert_true((
-            isset($config['packages']['git@github.com:saeghe/released-package.git'])
-            && 'v1.0.6' === $config['packages']['git@github.com:saeghe/released-package.git']
-        ),
-        $message
-    );
-}
-
-function assert_meta_updated($message)
-{
-    $meta = Json\to_array(realpath(root() . 'TestRequirements/Fixtures/EmptyProject/saeghe.config-lock.json'));
-
-    assert_true((
-            isset($meta['packages']['git@github.com:saeghe/released-package.git'])
-            && 'v1.0.6' === $meta['packages']['git@github.com:saeghe/released-package.git']['version']
-            && 'saeghe' === $meta['packages']['git@github.com:saeghe/released-package.git']['owner']
-            && 'released-package' === $meta['packages']['git@github.com:saeghe/released-package.git']['repo']
-            && '5885e5f3ed26c2289ceb2eeea1f108f7fbc10c01' === $meta['packages']['git@github.com:saeghe/released-package.git']['hash']
-        ),
-        $message
-    );
-}
-
-function assert_given_version_added($message)
-{
-    $config = Json\to_array(realpath(root() . 'TestRequirements/Fixtures/EmptyProject/saeghe.config.json'));
-    $meta = Json\to_array(realpath(root() . 'TestRequirements/Fixtures/EmptyProject/saeghe.config-lock.json'));
-
-    assert_true((
-            isset($config['packages']['git@github.com:saeghe/released-package.git'])
-            && 'v1.0.5' === $config['packages']['git@github.com:saeghe/released-package.git']
-            && isset($meta['packages']['git@github.com:saeghe/released-package.git'])
-            && 'v1.0.5' === $meta['packages']['git@github.com:saeghe/released-package.git']['version']
-            && 'saeghe' === $meta['packages']['git@github.com:saeghe/released-package.git']['owner']
-            && 'released-package' === $meta['packages']['git@github.com:saeghe/released-package.git']['repo']
-            && '5885e5f3ed26c2289ceb2eeea1f108f7fbc10c01' === $meta['packages']['git@github.com:saeghe/released-package.git']['hash']
-        ),
-        $message
-    );
-}
